+++ packages/core/core/src/types/runtime.ts
@@ -1,5 +1,5 @@
 import { IS_BUILT_IN } from '../symbols'
-import {
+import type {
   Event,
   Deserializer,
   ReadModel,
@@ -17,8 +17,13 @@ import {
   CommandContext,
   ReadModelResolverContext,
   ReadModelHandlerContext,
+  SerializableMap,
+  Command,
+  CommandResult,
 } from './core'
 import { CommandHttpResponseMode } from '../aggregate/types'
+import type { IncomingHttpHeaders } from 'http'
+import type { CookieSerializeOptions } from 'cookie'
 
 export type PerformanceSubsegment = {
   addAnnotation: (name: string, data: any) => void
@@ -37,28 +42,131 @@ export type Monitoring = {
   group: (config: Record<string, any>) => Monitoring
   error: (error: Error) => void
   execution: (error?: Error) => void
+  duration: (label: string, duration: number, count?: number) => void
   time: (name: string, timestamp?: number) => void
   timeEnd: (name: string, timestamp?: number) => void
   publish: () => Promise<void>
+  rate: (metricName: string, count: number, seconds?: number) => void
   performance?: PerformanceTracer
 }
 
+export type EventThreadData = {
+  threadCounter: number
+  threadId: number
+}
+
+export type StoredEvent = Event & EventThreadData
+
+export type Cursor = string
+export type InputCursor = string | null
+
+export type EventPointer = {
+  event: Event
+  cursor: Cursor
+}
+
+export type StoredEventPointer = {
+  cursor: Cursor
+  event: StoredEvent
+}
+
+export type StoredEventBatchPointer = {
+  cursor: Cursor
+  events: StoredEvent[]
+}
+
+type EventLimitFilter = {
+  limit: number
+  eventsSizeLimit?: number
+}
+type EventTypeFilter = {
+  aggregateIds?: string[] | null
+  eventTypes?: string[] | null
+}
+type EventTimestampFilter = {
+  startTime?: number
+  finishTime?: number
+}
+type EventCursorFilter = {
+  cursor?: InputCursor
+}
+
+export type EventFilter = EventLimitFilter &
+  EventTypeFilter &
+  EventTimestampFilter &
+  EventCursorFilter
+
+export type SecretRecord = {
+  idx: number
+  id: string
+  secret: string | null
+}
+
+export type OldSecretRecord = SecretRecord
+export type OldEvent = Event
+
+export type ReplicationStatus =
+  | 'batchInProgress'
+  | 'batchDone'
+  | 'error'
+  | 'notStarted'
+  | 'serviceError'
+export type ReplicationState = {
+  status: ReplicationStatus
+  statusData: SerializableMap | null
+  paused: boolean
+  iterator: SerializableMap | null
+  successEvent: OldEvent | null
+}
+
 export type Eventstore = {
-  saveEvent: (event: any) => Promise<any>
-  saveSnapshot: Function
-  getNextCursor: (cursor: any, events: Event[]) => Promise<any>
+  saveEvent: (event: Event) => Promise<StoredEventPointer>
+  saveSnapshot: (snapshotKey: string, content: string) => Promise<void>
+  getNextCursor: (cursor: InputCursor, events: EventThreadData[]) => string
   loadSnapshot: (snapshotKey: string) => Promise<string | null>
-  loadEvents: (param: {
-    aggregateIds: string[]
-    eventTypes?: string[]
-    cursor: null
-    limit: number
-  }) => Promise<{
-    events: any[]
-  }>
-  ensureEventSubscriber: Function
-  getEventSubscribers: Function
-  removeEventSubscriber: Function
+  loadEvents: (filter: EventFilter) => Promise<StoredEventBatchPointer>
+
+  ensureEventSubscriber: (params: {
+    applicationName: string
+    eventSubscriber: string
+    destination?: any
+    status?: any
+    updateOnly?: boolean
+  }) => Promise<boolean>
+  removeEventSubscriber: (params: {
+    applicationName: string
+    eventSubscriber: string
+  }) => Promise<void>
+  getEventSubscribers: (
+    params?:
+      | {
+          applicationName?: string
+          eventSubscriber?: string
+        }
+      | undefined
+  ) => Promise<
+    Array<{
+      applicationName: string
+      eventSubscriber: string
+      destination: any
+      status: any
+    }>
+  >
+
+  replicateEvents: (events: OldEvent[]) => Promise<void>
+  replicateSecrets: (
+    existingSecrets: OldSecretRecord[],
+    deletedSecrets: Array<OldSecretRecord['id']>
+  ) => Promise<void>
+  setReplicationIterator: (iterator: SerializableMap) => Promise<void>
+  setReplicationStatus: (
+    status: ReplicationStatus,
+    info?: ReplicationState['statusData'],
+    lastEvent?: OldEvent
+  ) => Promise<void>
+  setReplicationPaused: (pause: boolean) => Promise<void>
+  getReplicationState: () => Promise<ReplicationState>
+  resetReplication: () => Promise<void>
 }
 
 export type AggregateMeta = {
@@ -76,6 +184,7 @@ export type EventProjectionMeta = {
   name: string
   connectorName: string
   encryption: EventHandlerEncryptionFactory
+  invariantHash?: any
 }
 
 export type ReadModelMeta = EventProjectionMeta & {
@@ -99,6 +208,103 @@ export type ViewModelMeta = {
   encryption: EventHandlerEncryptionFactory
   invariantHash: string
 }
+
+export type Uploader = {
+  getSignedPut: (
+    dir: string
+  ) => Promise<{ uploadUrl: string; uploadId: string }>
+  getSignedPost: (dir: string) => Promise<{ form: any; uploadId: string }>
+  uploadPut: (uploadUrl: string, filePath: string) => Promise<void>
+  uploadPost: (form: { url: string }, filePath: string) => Promise<void>
+  createToken: (options: { dir: string; expireTime: number }) => string
+}
+
+export type BuildTimeConstants = {
+  applicationName: string
+  distDir: string
+  jwtCookie: {
+    name: string
+    maxAge: number
+  }
+  rootPath: string
+  staticDir: string
+  staticPath: string
+  staticRoutes?: string[] | undefined
+}
+
+export type QueryExecutor = {
+  (...args: any[]): Promise<any>
+  dispose: () => Promise<void>
+  [key: string]: (...args: any[]) => Promise<any>
+}
+
+export type SagaExecutor = QueryExecutor & {
+  build: (...args: any[]) => Promise<any>
+}
+
+export type CommandExecutor = {
+  (command: Command, context?: MiddlewareContext): Promise<CommandResult>
+  dispose: () => Promise<void>
+}
+
+export type UserBackendResolve = Readonly<
+  {
+    uploader: Uploader | null
+    eventstoreAdapter: Eventstore
+    performanceTracer: PerformanceTracer
+    executeCommand: CommandExecutor
+    executeQuery: QueryExecutor
+    executeSaga: SagaExecutor
+    seedClientEnvs: any
+    broadcastEvent: (event?: EventPointer) => Promise<void>
+  } & BuildTimeConstants
+>
+
+export type HttpRequest = {
+  readonly adapter: string
+  readonly method: string
+  readonly query: Record<string, any>
+  readonly path: string
+  readonly headers: IncomingHttpHeaders
+  readonly cookies: Record<string, string>
+  readonly body: string | null
+  jwt?: string
+  readonly isLambdaEdgeRequest?: boolean
+  readonly clientIp?: string
+}
+
+export type ResolveRequest = HttpRequest & {
+  readonly resolve: UserBackendResolve
+}
+
+export type HttpResponse = {
+  readonly cookie: (
+    name: string,
+    value: string,
+    options?: CookieSerializeOptions
+  ) => HttpResponse
+  readonly clearCookie: (
+    name: string,
+    options?: CookieSerializeOptions
+  ) => HttpResponse
+  readonly status: (code: number) => HttpResponse
+  readonly redirect: (path: string, code?: number) => HttpResponse
+  readonly getHeader: (searchKey: string) => any
+  readonly setHeader: (key: string, value: string) => HttpResponse
+  readonly text: (content: string, encoding?: BufferEncoding) => HttpResponse
+  readonly json: (content: any) => HttpResponse
+  readonly end: (
+    content?: string | Buffer,
+    encoding?: BufferEncoding
+  ) => HttpResponse
+  readonly file: (
+    content: string | Buffer,
+    filename: string,
+    encoding?: BufferEncoding
+  ) => HttpResponse
+}
+
+export type ResolveResponse = HttpResponse
 
 //Middleware
 
@@ -131,8 +337,8 @@ type MiddlewareChainableFunction =
 type MiddlewareHandler<THandler> = (next: THandler) => THandler
 
 export type MiddlewareContext = {
-  req?: any
-  res?: any
+  req?: ResolveRequest
+  res?: ResolveResponse
 }
 
 type ReadModelMiddlewareContext = {
@@ -166,3 +372,9 @@ export type ReadModelResolverMiddleware<
 export type CommandMiddleware<
   TContext extends CommandContext = CommandContext
 > = Middleware<CommandMiddlewareHandler<TContext>>
+
+export type ApiHandlerMeta = {
+  path: string
+  method: string
+  handler: (req: ResolveRequest, res: ResolveResponse) => Promise<void>
+}