+++ packages/runtime/runtime/src/__mocks__/resolve-cloud-common/postgres.js
@@ -1,3 +1,0 @@
-export const executeStatement = jest.fn().mockImplementation(() => [])
-export const escapeStr = jest.fn().mockImplementation((str) => str)
-export const escapeId = jest.fn().mockImplementation((str) => str)