+++ packages/runtime/runtime/src/common/defaults/upload-adapter.js
@@ -1,4 +1,0 @@
-// eslint-disable-next-line no-sequences
-const uploadAdapter = (options, imports) => (void imports, options)
-
-export default uploadAdapter