+++ packages/runtime/runtime/test/defaults/json-deserialize-state.test.js
@@ -1,6 +1,0 @@
-import { IS_BUILT_IN } from '@resolve-js/core'
-import deserialize from '../../src/common/defaults/json-deserialize-state'
-
-test('should have IS_BUILT_IN field set', () => {
-  expect(deserialize[IS_BUILT_IN]).toEqual(true)
-})