+++ packages/tools/scripts/src/alias/$resolve.clientEntry.js
@@ -4,7 +4,7 @@ import {
   RESOURCE_INSTANCE_ONLY,
   IMPORT_INSTANCE,
 } from '../constants'
-import importResource from '../import_resource'
+import { importResource } from '../import-resource'
 
 const importClientEntry = ({ resolveConfig, isClient }, resourceQuery) => {
   if (!/^\?/.test(resourceQuery)) {
@@ -28,9 +28,7 @@ const importClientEntry = ({ resolveConfig, isClient }, resourceQuery) => {
     : 'iife'
 
   const imports = [
-    `import getRootBasedUrl from '@resolve-js/runtime/lib/common/utils/get-root-based-url'`,
-    `import getStaticBasedPath from '@resolve-js/runtime/lib/common/utils/get-static-based-path'`,
-    `import jsonUtfStringify from '@resolve-js/runtime/lib/common/utils/json-utf-stringify'`,
+    `import { getRootBasedUrl, getStaticBasedPath, jsonUtfStringify }  from '@resolve-js/core'`,
   ]
 
   const constants = [