+++ packages/tools/scripts/src/alias/$resolve.staticPath.js
@@ -1,4 +1,4 @@
-import validatePath from '@resolve-js/runtime/lib/common/utils/validate-path'
+import { validatePath } from '@resolve-js/core'
 
 import { checkRuntimeEnv, injectRuntimeEnv } from '../declare_runtime_env'
 
@@ -21,9 +21,7 @@ const importStaticPath = ({ resolveConfig, isClient }) => {
     )
   } else {
     if (!isClient) {
-      exports.push(
-        `import validatePath from '@resolve-js/runtime/lib/common/utils/validate-path'`
-      )
+      exports.push(`import { validatePath } from '@resolve-js/core'`)
     }
 
     exports.push(`let staticPath = ${injectRuntimeEnv(staticPath, isClient)}`)