+++ templates/ts/postcss/config.app.ts
@@ -2,8 +2,10 @@ const appConfig = {
   apiHandlers: [
     {
       handler: {
-        module:
-          '@resolve-js/runtime/lib/common/handlers/live-require-handler.js',
+        module: {
+          package: '@resolve-js/runtime-base',
+          import: 'liveRequireHandler',
+        },
         options: {
           modulePath: './ssr.js',
           moduleFactoryImport: false,